requirements.txt
aiohttp
arxiv
beautifulsoup4
colorama
Flask
google
google-generativeai
GoogleNews
matplotlib
md2pdf
networkx
numpy
pdfplumber
Pillow
playwright
PyPDF2
requests
wikipedia
google-genai
docx